Create code
In this commit, the necessary libraries are installed and actions needed for data collection and data preparation are done to clean and prepare data for analysis and modeling.

As you can see the pandas object 'reads' the csv file in the directory and then the rows that have missing values are summed together via insull().sum(). 

Then, any rows that have missing values are removed. The data is then split into x and y features to suit machine  learning modelling. The data is split into training sets. And finally a print system is used to confirm that split training worked.